Reasonable implementation of HTTP/3.
This is the first iteration of Vert.x HTTP/3 implementation which implements partically RFC9114 and discovery via alternative services RFC7838

HTTP push is not implemented as we plan to drop this feature in Vert.x 6.

+++ vertx-core/src/main/java/io/vertx/core/http/impl/StatisticsGatheringHttpClientStream.java
@@ -38,7 +38,7 @@ class StatisticsGatheringHttpClientStream implements HttpClientStream {
   }
 
   @Override
-  public int id() {
+  public long id() {
     return delegate.id();
   }
